requirements.txt
colorama
dotenv
Pillow
requests
tqdm
Wand
streamlit